song-vocab/requirements.txt
pydantic
python-dotenv
streamlit
openai
openai-agents
pytest
pytest-asyncio